First version of Single-window library-style App
This changes make the application at least operational in a new mode called by Apple as "Single-window Library-style Application".
There are separate .xib files now for MainMenu and MainWindow, as it unclutters the code and permits the usage of more cleaner MVC classes as there is a standard App Delegate class as well as an isolated Window Controller class for the main window.
It's now working clearly with the new Xcode 5.