Begin change to remove additional ToolbarDelegate class
This patch is the initial work to integrate the V2BSMainLibraryToolbarDelegate into V2BSMainWindowController. It’s still a work in progress and not functional yet, but they are the first steps in that direction. There are yet a lot of rough corners to clean up (WIP), and the main window isn’t starting properly yet, so be warned!