Fix indentation problem project wide
This patch follows the lines of the additional fix of the previous commit. I performed a project wide search to fix all indentation changes done, even Xcode being configure to use spaces instead of tabs to indentate lines (even set to use spaces instead of tabs).
It seems that if follows the settings on new files, but keeps the old files untouched in this respect (even if you edited them with the new settings).